refactor: cleanup typespec
